requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.6.0
colorama==0.4.6
cryptography==41.0.7
distlib==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
filelock==3.13.1
#Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-secret-manager==2.17.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.2.0
mysqlclient==2.2.1
numpy==1.26.2
pandas==2.1.4
platformdirs==4.0.0
proto-plus==1.23.0
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.7
Werkzeug==3.0.1
WTForms==3.1.2
yarl==1.9.4
gunicorn==20.1.0